+++ src/pages/Articles/components/LostItemForm/LostItemForm.module.scss
@@ -101,31 +101,6 @@
   }
 }
 
-.title {
-  font-family: Pretendard, sans-serif;
-  font-size: 20px;
-  font-weight: 500;
-  line-height: 1.2;
-
-  @include media.media-breakpoint(mobile) {
-    font-size: 14px;
-    line-height: 1.6;
-  }
-
-  &__description {
-    color: #727272;
-    font-family: Pretendard, sans-serif;
-    font-size: 14px;
-    line-height: 1.6;
-
-    @include media.media-breakpoint(mobile) {
-      margin-bottom: 12px;
-      font-size: 12px;
-      line-height: 1.6;
-    }
-  }
-}
-
 .left {
   width: 559px;
   height: 254px;
@@ -147,97 +122,5 @@
 
   &--open {
     transform: rotate(180deg);
-  }
-}
-
-.warning {
-  position: absolute;
-  left: 0;
-  bottom: -32px;
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  color: #f7941e;
-  font-family: Pretendard, sans-serif;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 1.6;
-
-  @include media.media-breakpoint(mobile) {
-    left: initial;
-    right: 0;
-  }
-}
-
-.date {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-
-  @include media.media-breakpoint(mobile) {
-    width: 100%;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 8px;
-  }
-
-  &__wrapper {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    gap: 9px;
-
-    @include media.media-breakpoint(mobile) {
-      width: 100%;
-    }
-  }
-
-  &__toggle {
-    width: 332px;
-    height: 40px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    box-sizing: border-box;
-    padding: 7px 16px;
-    border-radius: 8px;
-    background-color: #f5f5f5;
-    cursor: pointer;
-
-    @include media.media-breakpoint(mobile) {
-      width: 100%;
-    }
-  }
-
-  &__description {
-    padding-top: 2px;
-    color: #727272;
-    font-family: Pretendard, sans-serif;
-    font-size: 14px;
-    font-weight: 400;
-    line-height: 1.6;
-
-    @include media.media-breakpoint(mobile) {
-      font-size: 12px;
-    }
-
-    &--has-been-selected {
-      color: #000;
-      font-size: 16px;
-      font-weight: 500;
-
-      @include media.media-breakpoint(mobile) {
-        font-size: 12px;
-      }
-    }
-  }
-
-  &__calendar {
-    position: absolute;
-    top: 48px;
-    right: 0;
-    z-index: 1;
   }
 }
